requirements.txt
aioredis==1.3.1
aiosqlite==0.15.0
anyio==3.7.1
asgiref==3.4.1
async-timeout==3.0.1
asyncpg==0.25.0
attrs==20.2.0
autobahn==21.3.1
Automat==20.2.0
cffi==1.14.2
channels==3.0.4
channels-redis==3.2.0
click==8.1.7
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
Django==3.2.13
django-sslserver==0.22
exceptiongroup==1.1.3
h11==0.14.0
hiredis==1.1.0
httptools==0.6.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==6.7.0
incremental==21.3.0
iso8601==0.1.13
msgpack==1.0.0
Pillow==9.0.1
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
PyPika==0.42.1
python-dotenv==0.19.2
pytz==2020.1
PyYAML==6.0.1
service-identity==18.1.0
six==1.15.0
sniffio==1.3.0
sqlparse==0.3.1
tortoise-orm==0.16.16
Twisted==22.4.0
txaio==21.2.1
typing-extensions==3.7.4.3
uvicorn==0.22.0
uvloop==0.18.0
watchfiles==0.20.0
websockets==11.0.3
zipp==3.15.0
zope.interface==5.1.0
